backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
pymongo==4.6.0

# Authentication and security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# NLP and ML libraries
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2

# Text processing and summarization
sumy==0.11.0
nltk==3.8.1

# Additional utilities
pydantic==2.5.0
python-dateutil==2.8.2
